[test] update bridged string tests
- move to a new file
- create a new `NSString` object for each new test
